improve key_to_index

+++ vlib/builtin/map.v
@@ -139,7 +139,7 @@ pub struct map {
 	// Byte size of value
 	value_bytes int
 mut:
-// Index of the highest index in the hashtable
+// highest even index in the hashtable
 	cap         u32
 	// Number of cached hashbits left for rehasing
 	window      byte
@@ -182,10 +182,10 @@ fn new_map_init(n, value_bytes int, keys &string, values voidptr) map {
 
 [inline]
 fn (m map) key_to_index(key string) (u32,u32) {
-	hash := u32(wyhash.wyhash_c(key.str, u64(key.len), 0))
+	hash := wyhash.wyhash_c(key.str, u64(key.len), 0)
 	index := hash & m.cap
 	meta := ((hash>>m.shift) & hash_mask) | probe_inc
-	return index,meta
+	return u32(index),u32(meta)
 }
 
 [inline]
